lobby et création de de joueur : fait Il faut maintenant synchroniser le debut des parties et gerer les actions des joueurs. Il faudra surement modifier ~~ completement la classe personnage pour que ce soit le joueur qui possède un personnage et non pas un personnage qui posssède un joueur

+++ src/game/elements/personnage/Personnage.hpp
@@ -41,7 +41,7 @@ public :
     void updateCollisionZone();
     void actions(sf::Event, bool allowingMovement);
 
-    bool owner(sf::Uint16 id);
+    bool owner(sf::Uint8 id);
     Player * getOwner();
     void Update(float dt);
     void updateAnimation();
